change toast, separate hook
