+++ public/assets/layout/staking-banner-bg.png
